add passport model score automation
upload_passport_model_scores.py reads the parquet data from a url provided by the passport team.
It then tries to create a temporary table (passport_model_scores_temp) in the grants db with this data.
If successful then it drops the existing passport_model_scores table and renames the temp table to not have the _temp suffix.
This approach should minimize downtime of the real table.

The corresponding .yml file runs every 4 hours or sooner if manually triggered